% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau.R
\name{coeff_tableau}
\alias{coeff_tableau}
\title{Extraction et formatage des coefficients d'un MCE}
\usage{
coeff_tableau(estim, divise_fr = TRUE)
}
\arguments{
\item{estim}{objet \code{lm} issu de l'estimation du MCE}

\item{divise_fr}{logique ; si \code{TRUE} (défaut), les coefficients de long terme
sont divisés par la force de rappel}
}
\value{
vecteur nommé de chaînes de caractères (HTML), un élément par variable
}
\description{
Combine coefficient, étoiles de significativité et t-statistique sous la forme
\code{"0.123*** <br>(2.45)"}. Pour les variables de long terme, le coefficient
est divisé par la force de rappel (élasticité de long terme). La force de rappel
reçoit des étoiles selon les valeurs critiques d'Ericsson-MacKinnon.
}
