% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delta}
\alias{delta}
\title{Lagged difference of order n}
\usage{
delta(n = 1, x)
}
\arguments{
\item{n}{lag order (default: 1)}

\item{x}{numeric vector}
}
\value{
vector of same length as x, with n leading NAs
}
\description{
Lagged difference of order n
}
