% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detrend}
\alias{detrend}
\title{Désaisonnalisation linéaire d'une série}
\usage{
detrend(serie)
}
\arguments{
\item{serie}{vecteur numérique}
}
\value{
résidus de la régression sur une tendance linéaire
}
\description{
Estime une tendance linéaire par MCO et retourne les résidus.
}
