% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcmOfce-package.R
\docType{package}
\name{EcmOfce-package}
\alias{EcmOfce}
\alias{EcmOfce-package}
\title{EcmOfce: Error Correction Model Tools for OFCE}
\description{
Fournit des outils pour l'estimation et la présentation des modèles à correction d'erreur (MCE) : tests diagnostics sur les résidus, valeurs critiques d'Ericsson-MacKinnon, simulations statiques et dynamiques, et mise en tableau des résultats.
}
\author{
\strong{Maintainer}: Elliot Aurissergues \email{elliot.aurissergues@gmail.com}

Authors:
\itemize{
  \item Elliot Aurissergues \email{elliot.aurissergues@gmail.com}
}

}
\keyword{internal}
