% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_plot_estim}
\alias{make_plot_estim}
\title{Graphiques résidus et simulation statique (version simple)}
\usage{
make_plot_estim(estim, data)
}
\arguments{
\item{estim}{objet \code{lm} issu de l'estimation}

\item{data}{data.frame contenant au moins une colonne \code{date}}
}
\value{
liste avec \code{plot_resid} et \code{plot_fit}
}
\description{
Représente la série observée vs la série simulée statiquement, et les résidus.
Les dates utilisées proviennent de la colonne \code{date} de \code{data}.
}
