% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_plot_estim2}
\alias{make_plot_estim2}
\title{Graphiques résidus et simulation statique (version interactive)}
\usage{
make_plot_estim2(estim, data)
}
\arguments{
\item{estim}{objet \code{lm} issu de l'estimation}

\item{data}{data.frame contenant au moins une colonne \code{date} et toutes
les variables de l'équation}
}
\value{
liste avec \code{plot_resid}, \code{plot_fit} (delta log) et
  \code{plot_fit_niv} (niveau)
}
\description{
Variante de \code{\link{make_plot_estim}} produisant des graphiques interactifs
via \pkg{ggiraph}, incluant un graphique en niveau et le thème OFCE.
}
