% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursif.R
\name{make_plot_recursif}
\alias{make_plot_recursif}
\title{Graphique d'estimation récursive}
\usage{
make_plot_recursif(
  data,
  model,
  start,
  end,
  date_debut_fixe = TRUE,
  plot_var_ct = FALSE,
  dummies = NULL
)
}
\arguments{
\item{data}{data.frame contenant les données}

\item{model}{objet \code{formula} du MCE}

\item{start}{première date de la séquence récursive}

\item{end}{dernière date de la séquence récursive}

\item{date_debut_fixe}{si \code{TRUE} (défaut), fait varier la date de fin}

\item{plot_var_ct}{si \code{TRUE}, affiche toutes les variables (hors dummies)}

\item{dummies}{vecteur d'indicatrices à gérer dynamiquement ; \code{NULL} par défaut}
}
\value{
objet \code{ggplot} interactif
}
\description{
Pour chaque variable du modèle, représente l'évolution du coefficient
selon la période d'estimation (date de fin ou de début variable).
Les t-statistiques sont affichées dans le tooltip interactif.
}
