% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delta}
\alias{delta}
\title{Différence d'ordre n d'une série}
\usage{
delta(n = 1, x)
}
\arguments{
\item{n}{ordre du décalage (défaut : 1)}

\item{x}{vecteur numérique}
}
\value{
vecteur de même longueur que x, avec n NA en tête
}
\description{
Différence d'ordre n d'une série
}
