% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detrend}
\alias{detrend}
\title{Linear detrending of a series}
\usage{
detrend(serie)
}
\arguments{
\item{serie}{numeric vector}
}
\value{
residuals from regression on a linear trend
}
\description{
Estimates a linear trend by OLS and returns the residuals.
}
